+++ src/cosi_db/controller/mod.rs
@@ -1,0 +1,2 @@
+pub mod dashboard;
+pub mod person;
