+++ src/cosi_db/mod.rs
@@ -2,3 +2,4 @@ pub mod common;
 pub mod connection;
 pub mod generator;
 pub mod person;
+pub mod controller;
